# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tags": {
      "get": {
        "description": "Requires admin or keeper privileges. Retrieves all live tag records (one per resource association) with optional filtering by resource type, search term, and sorting. Results are paginated - use Link and Pagination response headers for navigation.",
        "operationId": "TagController_findAll",
        "parameters": [
          {
            "name": "resourceType",
            "required": false,
            "in": "query",
            "description": "Filter tags by resource type (crux, path, or theme)",
            "schema": {
              "enum": [
                "crux",
                "path",
                "theme"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter tags by label (case-insensitive partial match)",
            "schema": {
              "example": "frontend",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort order: \"alpha\" for alphabetical, \"count\" for most used (default: count)",
            "schema": {
              "enum": [
                "alpha",
                "count"
              ],
              "type": "string"
            }
          },
          {
            "name": "label",
            "required": false,
            "in": "query",
            "description": "Filter by exact label match (case-sensitive)",
            "schema": {
              "example": "javascript",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination (default: 1)",
            "schema": {
              "example": 1
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "Number of results per page (default: 25, maximum: 100)",
            "schema": {
              "example": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tags retrieved successfully. Check Link header for pagination links and Pagination header for metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "tag_123"
                      },
                      "resourceType": {
                        "type": "string",
                        "enum": [
                          "crux",
                          "path",
                          "theme"
                        ],
                        "example": "crux"
                      },
                      "resourceId": {
                        "type": "string",
                        "example": "crux_456"
                      },
                      "label": {
                        "type": "string",
                        "example": "frontend-development"
                      },
                      "authorId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                      },
                      "homeId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                      },
                      "created": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updated": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all tags",
        "tags": [
          "Tags"
        ]
      }
    },
    "/tags/{id}": {
      "get": {
        "description": "Requires admin or keeper privileges. Retrieves a single tag by its UUID.",
        "operationId": "TagController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the tag",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tag retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "tag_123"
                    },
                    "resourceType": {
                      "type": "string",
                      "example": "crux"
                    },
                    "resourceId": {
                      "type": "string",
                      "example": "crux_456"
                    },
                    "label": {
                      "type": "string",
                      "example": "frontend-development"
                    },
                    "authorId": {
                      "type": "string",
                      "example": "author_789"
                    },
                    "created": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Tag not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a tag by ID",
        "tags": [
          "Tags"
        ]
      },
      "patch": {
        "description": "Updates the label of an existing tag by its ID. This is an admin-only operation and affects ALL resources using this tag.",
        "operationId": "TagController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the tag to update",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "tag_123"
                    },
                    "resourceType": {
                      "type": "string",
                      "enum": [
                        "crux",
                        "path",
                        "theme"
                      ],
                      "example": "crux"
                    },
                    "resourceId": {
                      "type": "string",
                      "example": "crux_456"
                    },
                    "label": {
                      "type": "string",
                      "example": "frontend-development",
                      "description": "Updated label (lowercase kebab-case)"
                    },
                    "authorId": {
                      "type": "string",
                      "example": "author_789"
                    },
                    "created": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid tag label format (must be lowercase kebab-case: a-z, 0-9, hyphens only)"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin permissions required"
          },
          "404": {
            "description": "Tag not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a tag",
        "tags": [
          "Tags"
        ]
      },
      "delete": {
        "description": "Soft deletes an existing tag by its ID. This is an admin-only operation and affects ALL resources using this tag. The tag is marked as deleted but not permanently removed from the database.",
        "operationId": "TagController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The UUID of the tag to delete",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Tag successfully deleted (soft delete). No content returned."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin permissions required"
          },
          "404": {
            "description": "Tag not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a tag",
        "tags": [
          "Tags"
        ]
      }
    },
...
    "/explore/reports": {
      "post": {
        "operationId": "ReportController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Report a published creation",
        "tags": [
          "explore"
        ]
      }
    },
    "/admin/reports": {
      "get": {
        "operationId": "ReportAdminController_reports",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "open",
                "resolved",
                "dismissed"
              ],
              "type": "string"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List reports, newest first",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/reports/{id}": {
      "patch": {
        "operationId": "ReportAdminController_updateReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Resolve, dismiss or reopen a report",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/takedowns": {
      "get": {
        "operationId": "ReportAdminController_takedowns",
        "parameters": [
          {
            "name": "active",
            "required": false,
            "in": "query",
            "description": "true = only takedowns still in force",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List takedowns, newest first",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "ReportAdminController_takeDown",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTakedownDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Take a crux down: unpublish it and refuse to publish that id again",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/takedowns/{cruxId}": {
      "delete": {
        "operationId": "ReportAdminController_liftTakedown",
        "parameters": [
          {
            "name": "cruxId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Lift a takedown; the author may publish the crux again",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Crux Garden API",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateTagDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateReportDto": {
        "type": "object",
        "properties": {
          "cruxId": {
            "type": "string",
            "description": "The published crux",
            "format": "uuid"
          },
          "reason": {
            "type": "string",
            "enum": [
              "illegal",
              "harmful",
              "spam",
              "copyright",
              "other"
            ]
          },
          "details": {
            "type": "string",
            "maxLength": 2000
          },
          "email": {
            "type": "string",
            "description": "Where the operator may reach the reporter",
            "format": "email"
          }
        },
        "required": [
          "cruxId",
          "reason"
        ]
      },
      "UpdateReportDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "resolved",
              "dismissed"
            ]
          },
          "resolutionNote": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateTakedownDto": {
        "type": "object",
        "properties": {
          "cruxId": {
            "type": "string",
            "description": "The crux to take down",
            "format": "uuid"
          },
          "reason": {
            "type": "string",
            "description": "Why, for the record",
            "maxLength": 2000
          },
          "reportId": {
            "type": "string",
            "description": "The report that prompted it",
            "format": "uuid"
          }
        },
        "required": [
          "cruxId",
          "reason"
        ]
      },
...
    }
  }
}